requirements.txt
django
pytest
psycopg2-binary
graphene-django
djangorestframework-jwt
djangorestframework